completed

+++ backend/routes/post.js
@@ -1,11 +1,15 @@
 const express = require ( 'express' );
-const router = express.Router();
+
 const checkAuth = require("../middleware/check-auth");
 const postmodel=require("../models/post");
-router.post("",(req, res, next)=>{
+
+const router = express.Router();
+
+router.post("", checkAuth, (req, res, next)=>{
   const post = new postmodel({
     title: req.body.title,
-    content: req.body.content
+    content: req.body.content,
+    creator: req.userData.userId
   });
   post.save().then(result=>{
     res.status(201).json({
@@ -16,22 +20,27 @@ router.post("",(req, res, next)=>{
   });
 });
 
-router.put("/:id", (req, res, next)=>{console.log('Inside PUT!')
+router.put("/:id", checkAuth, (req, res, next)=>{
   const post = new postmodel({
     _id: req.body.id,
     title: req.body.title,
-    content: req.body.content
+    content: req.body.content,
+    creator: req.userData.userId
   });
-  postmodel.updateOne({_id:req.params.id}, post).then(result =>{
-    console.log(result);
-    res.status(200).json({message: "Update Successful!"})
+  postmodel.updateOne({_id:req.params.id, creator: req.userData.userId}, post).then(result =>{
+    if(result.modifiedCount > 0){
+      //console.log('Inside PUT!:Update Successful!')
+      res.status(200).json({message: "Update Successful!"})
+    }else{
+      //console.log('Inside PUT!:Not Authorized!')
+      res.status(401).json({ message: "Not Authorized!" });
+    }
   });
 });
 
 router.get('', (req, res, next) =>{
   postmodel.find()
-  .then((documents)=>{console.log('Posts Found!')
-    //console.log(documents);
+  .then((documents)=>{
     res.status(200).json({
       message: 'Posts Fetched Successfully',
       posts: documents
@@ -42,23 +51,28 @@ router.get('', (req, res, next) =>{
 router.get("/:id",(req, res, next)=>{
   postmodel.findById(req.params.id).then(post =>{
     if(post){
-      console.log('Post Found!')
       res.status(200).json(post);
     }else{
-      console.log('Post not Found!')
+      //console.log('Post not Found!')
       res.status(484).json({message: 'Post not Found!'});
     }
   });
 });
 
-router.delete("/:id", (req, res, next)=>{
-  postmodel.deleteOne({_id:req.params.id}).then(result=>{
-    console.log(result);
-    res.status(200).json({
-      message:"Post deleted!"
-    });
+router.delete("/:id", checkAuth, (req, res, next)=>{
+  postmodel.deleteOne({_id:req.params.id,creator: req.userData.userId}).then(result=>{
+    if(result.deletedCount > 0){
+      res.status(200).json({ message: "Deleted successful!" });
+    }else{
+      res.status(401).json({ message: "Not Authorized!" });
+    }
   });
 });
+
+
+module.exports = router;
+
+
 /*router.post("",(req, res, next)=>{
   const post = new postmodel({
     title: req.body.title,
@@ -129,6 +143,8 @@ router.put("/:id", (req, res, next)=>{
     posts: posts
   });
   console.log('data service called')
-});*/
+});
 
 module.exports = router;
+*/
+
